refactor: GGFCharacterBase 클래스와 GGFCharacter 클래스 병합
